# dev-docs/openapi-mvp.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OSMU MVP API",
    "version": "0.1.0-prototype",
    "description": "Machine-readable contract snapshot for the current OSMU prototype. It covers the REST control plane and the MVP S3-compatible alias surface used by the local release gate."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local prototype backend"
    }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/admin/organizations": {
      "get": {
        "tags": ["Organizations", "Admin"],
        "operationId": "getOrganizations",
        "summary": "List organizations",
        "responses": {
          "200": { "$ref": "#/components/responses/JsonList" }
        }
      },
      "post": {
        "tags": ["Organizations", "Admin"],
        "operationId": "createOrganization",
        "summary": "Create an organization",
        "requestBody": { "$ref": "#/components/requestBodies/JsonObject" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "201": { "$ref": "#/components/responses/JsonOk" }
        }
      }
    },
...
    "/api/admin/organizations/{organizationId}": {
      "delete": {
        "tags": ["Organizations", "Admin"],
        "operationId": "deleteOrganization",
        "summary": "Delete an empty organization",
        "parameters": [
          { "$ref": "#/components/parameters/OrganizationId" }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/buckets/{bucketName}/objects": {
      "get": {
        "tags": ["Objects"],
        "operationId": "getObjects",
        "summary": "List, search, and filter objects",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" },
          { "$ref": "#/components/parameters/PrefixQuery" },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/TagQuery" },
          { "$ref": "#/components/parameters/DeletedQuery" },
          { "$ref": "#/components/parameters/LimitQuery" },
          { "$ref": "#/components/parameters/CursorQuery" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonList" }
        }
      },
      "post": {
        "tags": ["Objects"],
        "operationId": "uploadObject",
        "summary": "Upload one object through the REST API",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["key", "file"],
                "properties": {
                  "key": { "type": "string" },
                  "file": { "type": "string", "format": "binary" },
                  "tags": { "type": "string", "description": "Comma-separated key=value tags." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "201": { "$ref": "#/components/responses/JsonOk" }
        }
      }
    },
    "/api/buckets/{bucketName}/objects/{objectKey}": {
      "get": {
        "tags": ["Objects"],
        "operationId": "downloadObject",
        "summary": "Download an object",
        "description": "The Spring route captures a slash-containing object key.",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" },
          { "$ref": "#/components/parameters/ObjectKey" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/BinaryOk" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Objects"],
        "operationId": "deleteObject",
        "summary": "Soft-delete an object",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" },
          { "$ref": "#/components/parameters/ObjectKey" }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/buckets/{bucketName}/objects/share-links": {
      "post": {
        "tags": ["Objects"],
        "operationId": "createObjectShareLink",
        "summary": "Create a temporary object share link",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/JsonObject" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" }
        }
      },
      "get": {
        "tags": ["Objects"],
        "operationId": "getObjectShareLinks",
        "summary": "List temporary object share links for a bucket or object key",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" },
          {
            "name": "key",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonList" }
        }
      }
    },
...
    "/api/buckets/{bucketName}/objects/share-links/{linkId}": {
      "delete": {
        "tags": ["Objects"],
        "operationId": "deleteObjectShareLink",
        "summary": "Revoke a temporary object share link",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" },
          {
            "name": "linkId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "Revoked" }
        }
      }
    },
    "/api/public/share-links/{token}": {
      "get": {
        "tags": ["Objects"],
        "operationId": "downloadObjectShareLink",
        "summary": "Download an object through a temporary share token",
        "security": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "X-OSMU-Share-Password",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "Optional password for password-protected share links."
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Optional password for password-protected share links."
          },
          {
            "name": "X-Forwarded-For",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "Optional proxy client IP source for IP-restricted share links."
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/BinaryOk" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "osmuAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OSMU-Access-Key"
      },
      "awsSigV4": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "AWS Signature Version 4 authorization header or equivalent presigned query parameters."
      }
    },
    "parameters": {
      "BucketName": {
        "name": "bucketName",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$" }
      },
      "ObjectKey": {
        "name": "objectKey",
        "in": "path",
        "required": true,
        "description": "Slash-containing object keys are supported by Spring wildcard routes.",
        "schema": { "type": "string", "minLength": 1 }
      },
...
      "OrganizationId": {
        "name": "organizationId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
      "PrefixQuery": {
        "name": "prefix",
        "in": "query",
        "schema": { "type": "string" }
      },
      "SearchQuery": {
        "name": "search",
        "in": "query",
        "schema": { "type": "string" }
      },
      "TagQuery": {
        "name": "tag",
        "in": "query",
        "schema": { "type": "string" }
      },
      "DeletedQuery": {
        "name": "deleted",
        "in": "query",
        "schema": { "type": "boolean" }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 1000 }
      },
      "CursorQuery": {
        "name": "cursor",
        "in": "query",
        "schema": { "type": "string" }
      }
    },
    "requestBodies": {
      "JsonObject": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
...
    },
    "responses": {
      "JsonOk": {
        "description": "JSON response",
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
      "JsonList": {
        "description": "JSON list response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ListResponse" }
          }
        }
      },
...
      "BinaryOk": {
        "description": "Binary response",
        "content": {
          "application/octet-stream": {
            "schema": { "type": "string", "format": "binary" }
          }
        }
      },
      "Error": {
        "description": "OSMU JSON error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "ListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "type": "object", "additionalProperties": true }
          },
          "nextCursor": { "type": "string", "nullable": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "requestId": { "type": "string" }
        }
      }
    }
  }
}